On the books, it would be nice if for the authors we would be limited only to partners (object library.partner) that are defined as authors. Find a way to display those authors in a nice way to avoid a long list. We would like to improve the search capability on the books like base on title, ISBN or authors. It is possible to search on all those fields, but only with the advanced search. Only the name can be searched directly.
On the rentals list view, we would like to highlight the state of the rentals. For example it would be nice if the line in the list view were red when late. On the form view, a similar limitation to the one set on the authors on the books on the customer would be handy.

On the partner views, show the rentals only for partners that are customer as they are meaningless for authors and pollute the view. When on the list view of those partners, it would be nice to be able to filter the customers from the authors.